Add Exception telemetry, upgrade telemetry object to ITelemetryService2 (#314)
Fixes #246.

This adds a TraceListener which will respond to the events that StreamJsonRpc sends through its TraceSource. It checks for the specific trace event ID that the RPC library uses, and checks to make sure the arguments are sane. Adding more listeners or changing the trace level should be safe and won't break the listener.

Property naming is up to us, since it's a new event, but I mirrored some of what I saw in the extension, that is, camelCase names. The event name is the same as what the extension is using for its language server telemetry, but with a `/` afterward so we don't have any name conflicts and can do our own thing. I picked `rpc.exception` for this event, but it can be changed if wanted.