Log when interpreter is not found (#812)
For #809.

If this happens we probably need some better handling, but we should at least print something.

I made it an error, which steals focus as in #810. I can change it to info or similar, but the LS will likely fail to initialize when there is no interpreter (at the moment).