Send telemetry event on analysis complete (#876)
Closes #851.

Using the same definition of "complete" as the log message. If it did fewer than 100 items, then nothing is sent to prevent spamming for every small change. Open to suggestions for another arbitrary condition instead.

I'd include the system memory, but that'd require a bunch of code to make it work cross platform (or waiting for dotnet/corefx#22660).