Limit size of exact collections (#913)
Fixes #877.

Limit the size of exact collections to 1000 items (when they are created from the expressions). This is set high in order to not affect potentially large `__all__` definitions.

probablepeople has a dictionary with more than 150,000 key/value pairs, which takes a very long time to process; 1000 doesn't take very long.