addressed csharp compiler breaking changes around default with binary operators
more detail
https://github.com/dotnet/roslyn/blob/master/docs/compilers/CSharp/Compiler%20Breaking%20Changes%20-%20post%20VS2019.md

breaking change number 3

and here is excerpt of the change

...

dotnet/roslyn#35684 In C# 7.1 the resolution of a binary operator with a default literal could result in using an object equality and giving the literal a type object. For example, given a variable t of an unconstrained type T, t == default would be improperly allowed and emitted as t == default(object). Similarly, for a reference type without a custom == operator, x == default would be improperly allowed and emitted as x == default(object). In Visual Studio 2019 version 16.4 these scenarios will now produce an error.

....