implement AddImport (#1656)
* first draft of add import on python.

currently it does only basic behavior and works only for module known to path resolver and loaded modules (variable module)

no test yet. require many refactorings

* filter out unnecessary imports

* removed unused usings

* more refactoring

* addressed PR feedbacks

* added initial tests

* adding more tests

* rename

* simplify test a bit more

* added more tests

* changed codeaction source to use diagnostic service which should have diagnostic cached for latest analysis

* removed unused usings

* add ability to index library to IndexManager

* trying GetAnalysisAsync

* add closed module support for add imports.

* removed assert that keep crashing debug bit LS

the reason it was assert was due to code like below that defines 2 functions with same signature and we only keep 1 symbol out of it.
...
       def f(x):
            class C:
                x = 12
                def m(self):
                    return x
                locals()
            return C

        self.assertEqual(f(1).x, 12)

        def f(x):
            class C:
                y = x
                def m(self):
                    return x
                z = list(locals())
            return C
...

* add closed module test.

added just one since it is very expensive.

* use explicit type rather than tuple since it has many members

* add readonly

* use existing formatUI to format localizable string

* disabled locally imported code action for now

* fixed a bug where it didn't handle circular import properly, and added more comment

* handle circular imports better. some refactorings and etc.

* tweaked code action ordering a bit

* fix push/pop mismatch due to class decl being duplicated issue

* added test for symbol ordering

* added abbreviation for well known module

* added some comments

* added context based filtering

* clean up

* make sure we don't suggest imports that can be used in import statement

* added support for quick fix in hover

* added one more test and fixed one bug

* addressed easy feedback first

* made add import not to load modules implicitly and only use modules already loaded by users code.

for closed files that are never loaded in the editor before, it will use SymbolIndexer to find candidates.

but since the indexer is syntax based, it wont be able to provide information that require semantic info such as imported members

ex) from os import * or __all__.extend(core.__all__)

which means we will not propose

from os.path import join

if os module is not already loaded, instead we will propose ntpath or mac/posixpath since that is where join is actually defined.

unfortunately we can't load modules just for add import, it is too expensive.

we might in future add another index that is based on analyzed data which could have correct info saved. but for now, this is what we decided.

* added more abbreviation supports

* removed unnecessary time out

* support reverse abbreviation case where user types "np" and we suggest numpy

* made abbreviation to check existing symbol names to use unique name for abbreviation

* change order of where we show special one (abbreviation one).

now we show them at the top if there is any.

* add unit test for new ordering

* addressed PR feedbacks

* more refactoring