revert back .net standard to 2.0 since we are used in VS for PTVS which is still on .net fx
moved where we enable profile optimization back to the point where the server is initialized.

lang version is still left as latest even though it is "unsupported", according to .net people, it is common that people use latest with 2.0 and make sure some language features that doesnt work such as default interface implementation are not used.